refacto(admin) remove ununsed method
