+++ admin/app/controllers/authenticated/certification-centers/get.js
@@ -59,16 +59,6 @@ export default class AuthenticatedCertificationCentersGetController extends Cont
   }
 
   @action
-  updateGrantedAccreditation(accreditation) {
-    const accreditations = this.model.certificationCenter.accreditations;
-    if (accreditations.includes(accreditation)) {
-      accreditations.removeObject(accreditation);
-    } else {
-      accreditations.addObject(accreditation);
-    }
-  }
-
-  @action
   async updateCertificationCenter(certificationCenterData) {
     this.model.certificationCenter.name = certificationCenterData.name;
     this.model.certificationCenter.externalId = certificationCenterData.externalId;
