+++ admin/tests/unit/controllers/authenticated/certification-centers/get_test.js
@@ -203,52 +203,4 @@ module('Unit | Controller | authenticated/certification-centers/get', function (
       });
     });
   });
-
-  module('#updateGrantedAccreditation', function () {
-    test('it should add the accreditation to the certification center', function (assert) {
-      // given
-      const cleaAccreditation = { name: 'Pix+clea' };
-
-      certificationCenter = {
-        id: 2,
-        name: 'Centre des Anne-surfeuses',
-        type: 'PRO',
-        externalId: 'ex313',
-        accreditations: [],
-      };
-
-      controller.model = {
-        certificationCenter,
-      };
-
-      // when
-      controller.updateGrantedAccreditation(cleaAccreditation);
-
-      // then
-      assert.true(controller.model.certificationCenter.accreditations.includes(cleaAccreditation));
-    });
-
-    test('it should remove the accreditation from the certification center', function (assert) {
-      // given
-      const pixSurfAccreditation = { name: 'Pix+Surf' };
-
-      certificationCenter = {
-        id: 2,
-        name: 'Centre des Anne-surfeuses',
-        type: 'PRO',
-        externalId: 'ex313',
-        accreditations: [pixSurfAccreditation],
-      };
-
-      controller.model = {
-        certificationCenter,
-      };
-
-      // when
-      controller.updateGrantedAccreditation(pixSurfAccreditation);
-
-      // then
-      assert.false(controller.model.certificationCenter.accreditations.includes(pixSurfAccreditation));
-    });
-  });
 });
